+++ lib/td_dd/data_structures/error.exs
@@ -1,0 +1,10 @@
+TdDd.Loader.find_parent_child(%{
+  child_external_id:
+    "https://ll1qaeunestasrcsocial.blob.core.windows.net/source/followersbycountry/blinkfire//followersByCountry ########.csv",
+  child_group: "ll1qaeunestasrcsocial",
+  child_name: "followersByCountry ########.csv",
+  parent_external_id:
+    "https://ll1qaeunestasrcsocial.blob.core.windows.net/source/followersbycountry/blinkfire/",
+  parent_group: "ll1qaeunestasrcsocial",
+  system_id: 4
+})